+++ src/app/env.d.ts
@@ -1,0 +1,6 @@
+declare namespace NodeJS {
+  interface ProcessEnv {
+    OPENAI_API_KEY: string;
+    NODE_ENV: 'development' | 'production' | 'test';
+  }
+}
